refactor(http-server): flatten include directory
Move headers from inc/http_server/ → inc/.  All includes updated
from "http_server/xxx.hpp" to "xxx.hpp".  The CMakeLists.txt
include_directories already points at inc/ so no build change needed.
