+++ modules/http-server/src/router.cpp
@@ -1,4 +1,4 @@
-#include "http_server/router.hpp"
+#include "router.hpp"
 
 #include <sstream>
