+++ modules/http-server/src/worker.cpp
@@ -1,4 +1,4 @@
-#include "http_server/worker.hpp"
+#include "worker.hpp"
 
 namespace http_server {
